Lesson_1(client, server -> to classes)
