+++ .gitignore
@@ -1,2 +1,4 @@
 /Lesson_1_Yusupov_Idel.zip
 /env
+/.idea
+/Урок 2. Пример практического задания
