+++ Lessons/common/variables.py
